Feature: add string colorization
Defined a new string attribute for blob instances named :
"string_color".
This string (self.string_color) is going to be used only when drawing
the grid, to colorize blobs. The other one (self.string) is going to
be used for other things, such as computations for the length of
cells...

Add 50 to the randrange for color when generating a blob to avoid too
dark colors (not very due to the dark bg of the terminal)